# FamilyBenefits-api-openapi.yaml
openapi: 3.0.3
info:
  title: Family Benefits
  description: rest-api для приложения "Пособия для семьи с детьми"
  version: 1.0.0
  contact:
    email: smegovic@gmail.com
servers:
  - url: https://localhost
...
paths:
...
  /city:
    post:
      tags:
        - city
      summary: Добавляет новый город
      description:
        Добавляет новый город.
        Для администратора.
        Необходима авторизация.
      operationId: addCity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityAdd'
        required: true
      responses:
        '201':
          description: Город добавлен
          content: { }
        '400':
          description: Город с указанными данными уже существует
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    put:
      tags:
        - city
      summary: Обновляет город
      description:
        Обновляет город.
        Для администратора.
        Необходима авторизация.
      operationId: updateCity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityUpdate'
        required: true
      responses:
        '201':
          description: Город обновлен
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Город с указанными данными не найден
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
  /city/{id}:
    delete:
      tags:
        - city
      summary: Удаляет город
      description:
        Удаляет город.
        Для администратора.
        Необходима авторизация.
      operationId: deleteCity
      parameters:
        - name: id
          in: path
          description: ID города
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '201':
          description: Город удален
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Город с указанными данными не найден
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    get:
      tags:
        - city
      summary: Возвращает информацию о городе
      description:
        Возвращает информацию о городе.
      operationId: getCity
      parameters:
        - name: id
          in: path
          description: ID города
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '200':
          description: Данные о городе получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityInfo'
        '404':
          description: Город с указанным ID не найден
          content: { }
      security: [ ]
...
  /city/allpartial:
    get:
      tags:
        - city
      summary: Возвращает множество городов, в которых нет учреждений или пособий
      description:
        Возвращает множество городов, в которых нет учреждений или пособий.
        Для администратора.
        Необходима авторизация.
      operationId: getPartialCities
      responses:
        '200':
          description: Множество городов получено
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CityInfo'
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
  /city/initdata:
    get:
      tags:
        - city
      summary: Возвращает дополнительные данные для города
      description:
        Возвращает дополнительные данные для добавления или обновления города.
        Данные содержат в себе множество кратких информаций о пособиях.
      operationId: getCityInitData
      responses:
        '200':
          description: Дополнительные данные пользователя получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityInitData'
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
...
  /criterion:
    post:
      tags:
        - criterion
      summary: Добавляет новый критерий
      description:
        Добавляет новый критерий.
        Для администратора.
        Необходима авторизация.
      operationId: addCriterion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriterionAdd'
        required: true
      responses:
        '201':
          description: Критерий добавлен
          content: { }
        '400':
          description: Критерий с указанными данными уже существует
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    put:
      tags:
        - criterion
      summary: Обновляет данные критерия
      description:
        Обновляет данные критерия.
        Для администратора.
        Необходима авторизация.
      operationId: updateCriterion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriterionUpdate'
        required: true
      responses:
        '201':
          description: Критерий обновлен
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Не найден критерий с указанными данными
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
  /criterion/{id}:
    delete:
      tags:
        - criterion
      summary: Удаляет критерий
      description:
        Удаляет критерий.
        Для администратора.
        Необходима авторизация.
      operationId: deleteCriterion
      parameters:
        - name: id
          in: path
          description: ID критерия
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '201':
          description: Критерий удален
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Критерий с указанными данными не найден
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    get:
      tags:
        - criterion
      summary: Возвращает информацию о критерии
      description:
        Возвращает информацию о критерии.
      operationId: getCriterion
      parameters:
        - name: id
          in: path
          description: ID критерия
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '200':
          description: Данные о критерии получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CriterionInfo'
        '404':
          description: Критерий с указанным ID не найден
          content: { }
      security: [ ]
...
  /criterion/allpartial:
    get:
      tags:
        - criterion
      summary: Возвращает множество критерий, в которых нет пособий
      description:
        Возвращает множество критерий, в которых нет пособий.
        Для администратора.
        Необходима авторизация.
      operationId: getPartialCriteria
      responses:
        '200':
          description: Множество всех критерий получено
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CriterionInfo'
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
  /criterion/initdata:
    get:
      tags:
        - criterion
      summary: Возвращает дополнительные данные для критерия
      description:
        Возвращает дополнительные данные для добавления или обновления критерия.
        Данные содержат в себе множетсво кратких информаций о типах критерий.
        Для администратора.
        Необходима авторизация.
      operationId: getCriterionInitData
      responses:
        '200':
          description: Дополнительные данные критерия получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CriterionInitData'
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CriterionInfo:
      description: Информация о критерии для подбора пособий
      properties:
        id:
          description: ID критерия
          type: string
          example: td56387fhj
        name:
          description: Название критерия
          type: string
          example: name
        info:
          description: Информация критерия
          type: string
          example: info
        shortCriterionType:
          $ref: '#/components/schemas/ObjectShortInfo'
        shortBenefitSet:
          description: Множество кратких информаций о пособиях критерия
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
    CriterionAdd:
      description: Объект запроса для добавления критерия
      properties:
        name:
          description: Название критерия
          type: string
          example: name
        info:
          description: Информация критерия
          type: string
          example: info
        idCriterionType:
          description: ID типа критерия критерия
          type: string
          example: td56387fhj
    CriterionUpdate:
      description: Объект запроса для обновления критерия
      properties:
        id:
          description: ID критерия
          type: string
          example: td56387fhj
        name:
          description: Название критерия
          type: string
          example: name
        info:
          description: Информация критерия
          type: string
          example: info
        idCriterionType:
          description: ID типа критерия критерия
          type: string
          example: td56387fhj
    CriterionInitData:
      description: Дополнительные данные для добавления или обновления критерия
        Содержат в себе множество кратких информаций о типах критериях
      properties:
        shortCriterionTypeSet:
          description: Множество кратких информаций о типах критерия
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
    CityInfo:
      description: Информация о городе
      properties:
        id:
          description: ID города
          type: string
          example: td56387fhj
        name:
          description: Название города
          type: string
          example: name
        info:
          description: Информация города
          type: string
          example: info
        shortBenefitSet:
          description: Множество кратких информаций о пособиях города
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
        shortInstitutionSet:
          description: Множество кратких информаций об учреждениях города
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
    CityAdd:
      description: Объект запроса для добавления города
      properties:
        name:
          description: Название города
          type: string
          example: name
        info:
          description: Информация города
          type: string
          example: info
        idBenefitSet:
          description: Множество ID пособий города
          type: array
          items:
            description: ID пособия
            type: string
            example: td56387fhj
    CityUpdate:
      description: Объект запроса для обновления города
      properties:
        id:
          description: ID города
          type: string
          example: td56387fhj
        name:
          description: Название города
          type: string
          example: name
        info:
          description: Информация города
          type: string
          example: info
        idBenefitSet:
          description: Множество ID пособий города
          type: array
          items:
            description: ID пособия
            type: string
            example: td56387fhj
    CityInitData:
      description: Дополнительные данные для добавления или обновления города
        Содержат в себе множество кратких информаций о пособиях
      properties:
        shortBenefitSet:
          description: Множество кратких информаций о пособиях
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
...
    ObjectShortInfo:
      description: Краткая информация об объекте
      properties:
        idObject:
          description: ID объекта
          type: string
          example: td56387fhj
        nameObject:
          description: Название объекта
          type: string
          example: name